taking in rest of the comments
